update endpoint connected to cart page
