+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/viewmodels/CartViewModel.kt
@@ -8,7 +8,9 @@ import com.example.getflix.models.CartProductModel
 import com.example.getflix.models.CustomerCartPriceModel
 import com.example.getflix.models.ProductModel
 import com.example.getflix.service.GetflixApi
+import com.example.getflix.service.requests.CardProUpdateRequest
 import com.example.getflix.service.requests.CustomerCheckoutRequest
+import com.example.getflix.service.responses.CardProUpdateResponse
 import com.example.getflix.service.responses.CustomerCheckoutResponse
 import kotlinx.coroutines.*
 import retrofit2.Call
@@ -33,11 +35,12 @@ class CartViewModel : ViewModel() {
 
     fun getCustomerCartPrice() {
         job = CoroutineScope(Dispatchers.IO).launch {
-            val response = GetflixApi.getflixApiService.getCustomerCartPrice("Bearer " + MainActivity.StaticData.user!!.token)
+            val response =
+                GetflixApi.getflixApiService.getCustomerCartPrice("Bearer " + MainActivity.StaticData.user!!.token)
             withContext(Dispatchers.Main + exceptionHandler) {
                 if (response.isSuccessful) {
                     response.body().let { it ->
-                       _cardPrices.value = it
+                        _cardPrices.value = it
                     }
                 }
             }
@@ -46,18 +49,45 @@ class CartViewModel : ViewModel() {
 
     fun getCustomerCartProducts() {
         job = CoroutineScope(Dispatchers.IO).launch {
-            val response = GetflixApi.getflixApiService.getCustomerAllCartProducts("Bearer " + MainActivity.StaticData.user!!.token,
-                MainActivity.StaticData.user!!.id)
+            val response = GetflixApi.getflixApiService.getCustomerAllCartProducts(
+                "Bearer " + MainActivity.StaticData.user!!.token,
+                MainActivity.StaticData.user!!.id
+            )
             withContext(Dispatchers.Main + exceptionHandler) {
                 if (response.isSuccessful) {
                     println("succesfull mu")
                     response.body().let { it ->
-                        _cardProducts.value = response.body()!!.cartProducts as MutableList<CartProductModel>
+                        _cardProducts.value =
+                            response.body()!!.cartProducts as MutableList<CartProductModel>
                         println(_cardProducts.value.toString())
                     }
                 }
             }
         }
+    }
+
+    fun updateCustomerCartProduct(amount: Int, scId: Int, proId: Int) {
+        GetflixApi.getflixApiService.updateCustomerCartProduct(
+            "Bearer " + MainActivity.StaticData.user!!.token,
+            MainActivity.StaticData.user!!.id,
+            scId,
+            CardProUpdateRequest(proId, amount)
+        )
+            .enqueue(object :
+                Callback<CardProUpdateResponse> {
+                override fun onFailure(call: Call<CardProUpdateResponse>, t: Throwable) {
+
+                }
+
+                override fun onResponse(
+                    call: Call<CardProUpdateResponse>,
+                    response: Response<CardProUpdateResponse>
+                ) {
+                    getCustomerCartProducts()
+                    getCustomerCartPrice()
+                }
+            }
+            )
     }
